update the fixing
The real problem is that the "implementation" is actually template, which didn't compile them into real function when compiling matrix.cpp (there are others). Only when the compiler reads code that call the function which specifies the type in the template, the compiler will output the function. This is why the commit before the last one always report "undefined refrence to ..."